+++ src/index.js
@@ -1,18 +1,11 @@
-import gameWorld from './world/config';
+import gameStart from './gamestart';
 import './styles.css';
 
-const display = document.getElementById('gameStart');
-display.addEventListener('click', () => {
-  const playerName = document.getElementById('playerName').value;
+const index = () => {
 
-  if (playerName.length >= 3) {
-    localStorage.setItem('player', playerName);
-    localStorage.setItem('score', 0);
-    document.body.innerHTML = '';
-    gameWorld();
-  } else {
-    const warning = document.createElement('small');
-    warning.innerHTML = 'The name must me at least 3 characters long!!!';
-    document.getElementById('playerNameContainer').appendChild(warning);
-  }
-});
+  gameStart();
+
+}
+
+index();
+
