--- src/scoreboard/create_game.js
@@ -1,20 +1,0 @@
-const createGame = () => {
-  const gatherData = async () => {
-    const result = await fetch('https://us-central1-js-capstone-backend.cloudfunctions.net/api/games', {
-      method: 'POST',
-      mode: 'cors',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      redirect: 'follow',
-      body: JSON.stringify({ name: 'TadeuDragonSlayer' }),
-
-    });
-
-    return (result.json());
-  };
-
-  return (gatherData());
-};
-
-export default createGame;
